perf(bench): split query bench binary into rows/procs/prepared
The query bench binary was the largest (9 benches) and its within-binary
candidate->baseline time gap produced the residual base-slower drift on the
query benches (stored_procedure ~+18%) even on identical code. Splitting it
into three ~3-bench binaries (query_rows, query_procs, query_prepared) shrinks
each binary's span so the perf-lab's per-binary interleaving keeps each
benchmark's two measurements close together in time. Benchmark ids are
unchanged, so critcmp baseline names and comparison rows are preserved.
